+++ src/main/java/com/server/fabula/Model/Request/SignUpRequest.java
@@ -1,19 +1,15 @@
-package com.server.fabula.DAO.Request;
+package com.server.fabula.Model.Request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class UpdateUserRequest {
-
-    private Integer id;
+public class SignUpRequest {
     private String name;
     private String email;
-    private String currentPassword;
-    private String newPassword;
+    private String password;
 }
