+++ src/main/java/com/server/fabula/Model/Request/UpdateUserRequest.java
@@ -1,15 +1,19 @@
-package com.server.fabula.DAO.Request;
+package com.server.fabula.Model.Request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class SignUpRequest {
+public class UpdateUserRequest {
+
+    private Integer id;
     private String name;
     private String email;
-    private String password;
+    private String currentPassword;
+    private String newPassword;
 }
